test(api-mailer): simplify creating handlers
